+++ packages/client/src/actions/index.ts
@@ -1,6 +1,5 @@
 export * from './account';
 export * from './activity';
-export * from './activity-v2';
 export * from './approvals';
 export * from './auth';
 export * from './builders';
